2015-09-25 11:41 UTC+0200 Przemyslaw Czerpak (druzus/at/poczta.onet.pl)   * src/rtl/filebuf.c     ! removed unnecessary HB_EXPORT attribute
  * contrib/hbwin/oleuuid.c
    ! added missing HB_EXPORT attrinute for OpenWatcom builds

  * contrib/hbcomio/comio.c
  * contrib/hbpipeio/pipeio.c
  * contrib/hbtcpio/tcpio.c
  * src/rtl/hbsockhb.c
    ! cleaned integer types and casting

+++ src/rtl/hbsockhb.c
@@ -784,21 +784,21 @@ void hb_socekxParamsGetStd( PHB_ITEM pParams,
           HB_IS_STRING( pItem ) )
       {
          *pKeydata = hb_itemGetCPtr( pItem );
-         *pKeylen  = hb_itemGetCLen( pItem );
+         *pKeylen  = ( int ) hb_itemGetCLen( pItem );
       }
       else if( pKeydata && pKeylen &&
                ( pItem = hb_hashGetCItemPtr( pParams, "pass" ) ) != NULL &&
                HB_IS_STRING( pItem ) )
       {
          *pKeydata = hb_itemGetCPtr( pItem );
-         *pKeylen  = hb_itemGetCLen( pItem );
+         *pKeylen  = ( int ) hb_itemGetCLen( pItem );
       }
       if( pIV && pIVlen &&
           ( pItem = hb_hashGetCItemPtr( pParams, "iv" ) ) != NULL &&
           HB_IS_STRING( pItem ) )
       {
          *pIV    = hb_itemGetCPtr( pItem );
-         *pIVlen = hb_itemGetCLen( pItem );
+         *pIVlen = ( int ) hb_itemGetCLen( pItem );
       }
       if( pLevel &&
           ( pItem = hb_hashGetCItemPtr( pParams, "zlib" ) ) != NULL &&
